# web/openapi.json
{
  "components": {
    "schemas": {
      "AliasOut": {
        "properties": {
          "seen_at": {
            "title": "Seen At",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/Source"
          },
          "source_job_id": {
            "title": "Source Job Id",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "source",
          "source_job_id",
          "url",
          "seen_at"
        ],
        "title": "AliasOut",
        "type": "object"
      },
      "ApplicationStatus": {
        "description": "Where an application stands \u2014 the user's, set by hand. Absent means \"not tracked\".",
        "enum": [
          "applied",
          "interview",
          "offer",
          "rejected",
          "withdrawn"
        ],
        "title": "ApplicationStatus",
        "type": "string"
      },
...
      "CompensationOut": {
        "properties": {
          "amount_max": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Max"
          },
          "amount_min": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Min"
          },
          "bonus_mentioned": {
            "title": "Bonus Mentioned",
            "type": "boolean"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "equity_mentioned": {
            "title": "Equity Mentioned",
            "type": "boolean"
          },
          "period": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SalaryPeriod"
              },
              {
                "type": "null"
              }
            ]
          },
          "raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw"
          }
        },
        "required": [
          "amount_min",
          "amount_max",
          "currency",
          "period",
          "bonus_mentioned",
          "equity_mentioned",
          "raw"
        ],
        "title": "CompensationOut",
        "type": "object"
      },
...
      "FlagRequest": {
        "properties": {
          "value": {
            "title": "Value",
            "type": "boolean"
          }
        },
        "required": [
          "value"
        ],
        "title": "FlagRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LocationOut": {
        "properties": {
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "raw": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "remote_mode": {
            "$ref": "#/components/schemas/RemoteMode"
          }
        },
        "required": [
          "city",
          "country",
          "region",
          "remote_mode",
          "raw"
        ],
        "title": "LocationOut",
        "type": "object"
      },
...
      "NoteRequest": {
        "properties": {
          "note": {
            "maxLength": 4000,
            "title": "Note",
            "type": "string"
          }
        },
        "required": [
          "note"
        ],
        "title": "NoteRequest",
        "type": "object"
      },
      "PostingDetailOut": {
        "properties": {
          "alias_count": {
            "title": "Alias Count",
            "type": "integer"
          },
          "aliases": {
            "items": {
              "$ref": "#/components/schemas/AliasOut"
            },
            "title": "Aliases",
            "type": "array"
          },
          "application_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApplicationStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "closes_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closes At"
          },
          "company_name": {
            "title": "Company Name",
            "type": "string"
          },
          "company_slug": {
            "title": "Company Slug",
            "type": "string"
          },
          "compensation": {
            "$ref": "#/components/schemas/CompensationOut"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "favorited": {
            "title": "Favorited",
            "type": "boolean"
          },
          "first_seen_at": {
            "title": "First Seen At",
            "type": "string"
          },
          "hidden": {
            "title": "Hidden",
            "type": "boolean"
          },
          "last_seen_at": {
            "title": "Last Seen At",
            "type": "string"
          },
          "locations": {
            "items": {
              "$ref": "#/components/schemas/LocationOut"
            },
            "title": "Locations",
            "type": "array"
          },
          "min_years": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Years"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "phd_required": {
            "title": "Phd Required",
            "type": "boolean"
          },
          "posted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posted At"
          },
          "posting_id": {
            "title": "Posting Id",
            "type": "string"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/Reason"
            },
            "title": "Reasons",
            "type": "array"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason"
          },
          "role_family": {
            "$ref": "#/components/schemas/RoleFamily"
          },
          "score": {
            "title": "Score",
            "type": "integer"
          },
          "sector": {
            "title": "Sector",
            "type": "string"
          },
          "seniority": {
            "$ref": "#/components/schemas/Seniority"
          },
          "source": {
            "$ref": "#/components/schemas/Source"
          },
          "tech": {
            "items": {
              "type": "string"
            },
            "title": "Tech",
            "type": "array",
            "uniqueItems": true
          },
          "tier": {
            "$ref": "#/components/schemas/Tier"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "title_raw": {
            "title": "Title Raw",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "visa_sponsorship": {
            "$ref": "#/components/schemas/VisaStatus"
          }
        },
        "required": [
          "posting_id",
          "company_slug",
          "company_name",
          "sector",
          "source",
          "url",
          "title",
          "title_raw",
          "role_family",
          "seniority",
          "min_years",
          "phd_required",
          "locations",
          "compensation",
          "visa_sponsorship",
          "tech",
          "posted_at",
          "first_seen_at",
          "last_seen_at",
          "closes_at",
          "score",
          "tier",
          "alias_count",
          "favorited",
          "hidden",
          "application_status",
          "description",
          "note",
          "reasons",
          "rejection_reason",
          "aliases"
        ],
        "title": "PostingDetailOut",
        "type": "object"
      },
      "PostingOut": {
        "properties": {
          "alias_count": {
            "title": "Alias Count",
            "type": "integer"
          },
          "application_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApplicationStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "closes_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closes At"
          },
          "company_name": {
            "title": "Company Name",
            "type": "string"
          },
          "company_slug": {
            "title": "Company Slug",
            "type": "string"
          },
          "compensation": {
            "$ref": "#/components/schemas/CompensationOut"
          },
          "favorited": {
            "title": "Favorited",
            "type": "boolean"
          },
          "first_seen_at": {
            "title": "First Seen At",
            "type": "string"
          },
          "hidden": {
            "title": "Hidden",
            "type": "boolean"
          },
          "last_seen_at": {
            "title": "Last Seen At",
            "type": "string"
          },
          "locations": {
            "items": {
              "$ref": "#/components/schemas/LocationOut"
            },
            "title": "Locations",
            "type": "array"
          },
          "min_years": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Years"
          },
          "phd_required": {
            "title": "Phd Required",
            "type": "boolean"
          },
          "posted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posted At"
          },
          "posting_id": {
            "title": "Posting Id",
            "type": "string"
          },
          "role_family": {
            "$ref": "#/components/schemas/RoleFamily"
          },
          "score": {
            "title": "Score",
            "type": "integer"
          },
          "sector": {
            "title": "Sector",
            "type": "string"
          },
          "seniority": {
            "$ref": "#/components/schemas/Seniority"
          },
          "source": {
            "$ref": "#/components/schemas/Source"
          },
          "tech": {
            "items": {
              "type": "string"
            },
            "title": "Tech",
            "type": "array",
            "uniqueItems": true
          },
          "tier": {
            "$ref": "#/components/schemas/Tier"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "title_raw": {
            "title": "Title Raw",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "visa_sponsorship": {
            "$ref": "#/components/schemas/VisaStatus"
          }
        },
        "required": [
          "posting_id",
          "company_slug",
          "company_name",
          "sector",
          "source",
          "url",
          "title",
          "title_raw",
          "role_family",
          "seniority",
          "min_years",
          "phd_required",
          "locations",
          "compensation",
          "visa_sponsorship",
          "tech",
          "posted_at",
          "first_seen_at",
          "last_seen_at",
          "closes_at",
          "score",
          "tier",
          "alias_count",
          "favorited",
          "hidden",
          "application_status"
        ],
        "title": "PostingOut",
        "type": "object"
      },
      "PostingsPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PostingOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "PostingsPage",
        "type": "object"
      },
...
      "RemoteMode": {
        "enum": [
          "onsite",
          "hybrid",
          "remote",
          "unknown"
        ],
        "title": "RemoteMode",
        "type": "string"
      },
      "RoleFamily": {
        "enum": [
          "quant_dev",
          "quant_research",
          "quant_trading",
          "swe_platform",
          "data_eng",
          "risk",
          "other"
        ],
        "title": "RoleFamily",
        "type": "string"
      },
...
      "Seniority": {
        "enum": [
          "intern",
          "graduate",
          "junior",
          "mid",
          "senior",
          "lead",
          "unknown"
        ],
        "title": "Seniority",
        "type": "string"
      },
      "SortKey": {
        "enum": [
          "score",
          "posted",
          "seen",
          "company",
          "closes"
        ],
        "title": "SortKey",
        "type": "string"
      },
      "Source": {
        "enum": [
          "greenhouse",
          "lever",
          "ashby",
          "smartrecruiters",
          "workable",
          "recruitee",
          "personio",
          "workday",
          "custom",
          "efinancialcareers",
          "wttj",
          "linkedin",
          "indeed",
          "adzuna"
        ],
        "title": "Source",
        "type": "string"
      },
...
      "StatusRequest": {
        "description": "`status: null` stops tracking the posting.",
        "properties": {
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApplicationStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "status"
        ],
        "title": "StatusRequest",
        "type": "object"
      },
      "Tier": {
        "enum": [
          "strong",
          "possible",
          "stretch",
          "rejected"
        ],
        "title": "Tier",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VisaStatus": {
        "enum": [
          "sponsors",
          "no",
          "unknown"
        ],
        "title": "VisaStatus",
        "type": "string"
      }
    }
  },
  "info": {
    "title": "JobTracker API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/postings": {
      "get": {
        "operationId": "get_postings_postings_get",
        "parameters": [
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortKey",
              "default": "score"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "countries",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "type": "string"
              },
              "title": "Countries",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "cities",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "type": "string"
              },
              "title": "Cities",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "statuses",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "$ref": "#/components/schemas/ApplicationStatus"
              },
              "title": "Statuses",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "place_country",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Place Country"
            }
          },
          {
            "in": "query",
            "name": "place_city",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Place City"
            }
          },
          {
            "in": "query",
            "name": "companies",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "type": "string"
              },
              "title": "Companies",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "sectors",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "type": "string"
              },
              "title": "Sectors",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "sources",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "$ref": "#/components/schemas/Source"
              },
              "title": "Sources",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "role_families",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "$ref": "#/components/schemas/RoleFamily"
              },
              "title": "Role Families",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "seniorities",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "$ref": "#/components/schemas/Seniority"
              },
              "title": "Seniorities",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "remote_modes",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "$ref": "#/components/schemas/RemoteMode"
              },
              "title": "Remote Modes",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "tech_all",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "type": "string"
              },
              "title": "Tech All",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "tech_any",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "type": "string"
              },
              "title": "Tech Any",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "visa",
            "required": false,
            "schema": {
              "default": [
                "sponsors",
                "unknown"
              ],
              "items": {
                "$ref": "#/components/schemas/VisaStatus"
              },
              "title": "Visa",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "min_score",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 100,
              "minimum": 0,
              "title": "Min Score",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "tiers",
            "required": false,
            "schema": {
              "default": [
                "strong",
                "possible",
                "stretch"
              ],
              "items": {
                "$ref": "#/components/schemas/Tier"
              },
              "title": "Tiers",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "posted_within_days",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Posted Within Days"
            }
          },
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          },
          {
            "in": "query",
            "name": "favorites_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Favorites Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_hidden",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Hidden",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostingsPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Postings",
        "tags": [
          "postings"
        ]
      }
    },
    "/postings/{posting_id}": {
      "get": {
        "operationId": "get_posting_detail_postings__posting_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "posting_id",
            "required": true,
            "schema": {
              "title": "Posting Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostingDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Posting Detail",
        "tags": [
          "postings"
        ]
      }
    },
    "/postings/{posting_id}/favorite": {
      "post": {
        "operationId": "set_favorite_postings__posting_id__favorite_post",
        "parameters": [
          {
            "in": "path",
            "name": "posting_id",
            "required": true,
            "schema": {
              "title": "Posting Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlagRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Favorite",
        "tags": [
          "postings"
        ]
      }
    },
    "/postings/{posting_id}/hide": {
      "post": {
        "operationId": "set_hidden_postings__posting_id__hide_post",
        "parameters": [
          {
            "in": "path",
            "name": "posting_id",
            "required": true,
            "schema": {
              "title": "Posting Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlagRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Hidden",
        "tags": [
          "postings"
        ]
      }
    },
    "/postings/{posting_id}/note": {
      "post": {
        "operationId": "save_note_postings__posting_id__note_post",
        "parameters": [
          {
            "in": "path",
            "name": "posting_id",
            "required": true,
            "schema": {
              "title": "Posting Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Note",
        "tags": [
          "postings"
        ]
      }
    },
    "/postings/{posting_id}/status": {
      "post": {
        "operationId": "set_status_postings__posting_id__status_post",
        "parameters": [
          {
            "in": "path",
            "name": "posting_id",
            "required": true,
            "schema": {
              "title": "Posting Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Status",
        "tags": [
          "postings"
        ]
      }
    }
  }
}